app/requirements.txt
# Основной фреймворк для создания API
fastapi==0.109.0

# ASGI-сервер для запуска FastAPI
uvicorn==0.27.0

# ORM для работы с базой данных
sqlalchemy==2.0.25

# Драйвер PostgreSQL для Python
psycopg[binary]>=3.1.0

# Загрузка переменных окружения из .env файла
python-dotenv==1.0.0

# Валидация данных (используется внутри FastAPI)
pydantic>=2.8